Feature Request #51: Undo popup after deleting item
Done some additional work today.

When you delete an item from the list, the app will show the popup at the bottom of the screen. If you delete multiples, it will stack the popups on top of each other.

Bit more work to go however:

Right now for some reason the first popup doesn't stack at the bottom (It seems to think it's in position 2, when it needs to be in position 1.)
Deleted item needs to be shown in the popup so the user knows what they want to undo
POST Command to the REST API by ID to set the Active Flag back to 1.
Implement MVVM in the application.
Thats all I can think of right now.

See:
http://bugs.notsoslow.local/issues/51

Signed-off-by: Matthew Jones <jonesmatt1991@gmail.com>